Refactor `NCMenuAction` select actions
Generalise with static func as named initializers, to avoid code duplication.
Use `selectActions` as customisable option for `NCTrash`. Also avoids to use defer.
